New start in this version

+++ lvgl/tests/src/test_cases/test_demo_stress.c
@@ -1,11 +1,9 @@
 #if LV_BUILD_TEST
 #include "../lvgl.h"
+#include "../../lvgl_private.h"
 #include "../demos/lv_demos.h"
 
 #include "unity/unity.h"
-
-#include "lv_test_helpers.h"
-#include "lv_test_indev.h"
 
 static void loop_through_stress_test(void)
 {
@@ -26,7 +24,7 @@ void test_demo_stress(void)
 #endif
     /* loop once to allow objects to be created */
     loop_through_stress_test();
-    uint32_t mem_before = lv_test_get_free_mem();
+    size_t mem_before = lv_test_get_free_mem();
     /* loop 5 more times */
     for(uint32_t i = 0; i < 5; i++) {
         loop_through_stress_test();
